Update glyph classes
